HACK: staging: ion: Try to cache mappings to avoid cache invalidations
Since 4.12, much later narrowed down to commit 2a55e7b5e544
("staging: android: ion: Call dma_map_sg for syncing and mapping"),
we have seen graphics performance issues on the HiKey960.

After chasing a number of other leads, I found that reverting
the ION code to 4.11-era got the majority of the graphics
performance back (there may yet be further EAS tweaks needed),
which lead me to the dma_map_sg change.

In talking w/ Laura and Liam, it was suspected that the extra
cache operations were causing the trouble. Additionally, I found
that part of the reason we didn't see this w/ the original
HiKey board is that its (proprietary blob) GL code uses ion_mmap
and ion_map_dma_buf is called very rarely, where as with
HiKey960, the (also proprietary blob) GL code calls
ion_map_dma_buf much more frequently via the kernel driver.

Anyway, with the cause of the performance regression isolated,
I've tried to find a way to improve the performance of the
current code.

This approach, which reworks my previous attempt, tries to
use the dma-attributes to avoid doing cache operations if
the current direction is the same.

NOTE: Technically, this probably isn't the right solution for
every ION user, thus the HACK prefix. However, for HiKey960 it
resolves a pretty big interactive performance issue, and I
think we should consider merging it so we can switch to 4.14
as the default kernel in the near term. Once a proper fix is
found upstream, we can backport it.

Change-Id: Ic7ce3dde5efe4c4dc7a1f3cdf138f0c572a2ba35
Signed-off-by: John Stultz <john.stultz@linaro.org>