thermal/drivers/hisi_thermal: Fix merge issue coming from linux-4.14
The commits:
 dd9989eda8321 thermal/drivers/hisi: Remove costly sensor inspection
 aa6e035f259a2 thermal/drivers/hisi: Fix configuration register setting
 f7105bd037556 thermal/drivers/hisi: Encapsulate register writes into helpers
 e29649e49c81f thermal/drivers/hisi: Remove pointless lock
 402519439ccb4 thermal/drivers/hisi: Remove the multiple sensors support

were merged in linux-4.14 in order to solve a kernel WARNING spotted by
the lkft (https://lkft.linaro.org/). After merging android-4.14 and linux-4.14,
it appears there were conflicts on the driver. This conflict should have
not happen. It appeared these patches above and more changes on the hikey
thermal driver were merged through the ACK.

In order to fix the situation, we reverted the changes coming from ACK where
the hikey specific changes were undone, thus reimporting the changes in the
correct branch.

However, there is still a pending conflict with the 'setup' function left.

Remove it.

With this patch applied and the previous one, we end up with the same driver
as the one before the latest merge with ACK.

Change-Id: I0d5ea6537928cda7ee806b15ffae4a3131efd51a
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>